Increase privacy and clarity in registration form
The form that appears as a result of the "register" command now assigns type attribute to inputs of the registration form. The password inputs are given "type='password'" for greater privacy, and the userName input is given "type='input'" to more strongly bind that function to that element, and for greater clarity. 
A comment was added to existing code related to this alteration.